THE_LAST_DINOSAUR/requirements.txt
+certifi==2021.10.8
charset-normalizer==2.0.9
decorator==5.1.0
docopt==0.6.2
idna==3.3
numpy==1.21.2
Pillow==8.4.0
pipreqs==0.4.11
pygame==2.0.1
PyYAML==5.4.1
requests==2.26.0
self==2020.12.3
turtle==0.0.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
